Bonus Challenge done + extras

+++ Prototype 5/Assets/Scripts/DifficultyButton.cs
@@ -7,7 +7,7 @@ public class DifficultyButton : MonoBehaviour
 {
     private Button button;
     private GameManager gameManager;
-    public int difficulty;
+    public float difficulty;
 
     // Start is called before the first frame update
     void Start()
@@ -21,7 +21,6 @@ public class DifficultyButton : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        
     }
     void SetDifficulty()
     {
